modified:   .github/workflows/go-test.yaml 	new file:   api/schemas/browser.go 	new file:   api/schemas/common.go 	new file:   api/schemas/findings.go 	new file:   api/schemas/har.go 	new file:   api/schemas/interfaces.go 	new file:   api/schemas/knowledge_graph.go 	deleted:    api/schemas/schemas.go 	new file:   api/schemas/tasks.go 	modified:   cmd/evolution.go 	modified:   cmd/evolution_test.go 	modified:   cmd/main_test.go 	modified:   cmd/root.go 	modified:   cmd/scalpel/main.go 	modified:   cmd/scan.go 	modified:   cmd/scan_test.go 	modified:   cmd/self_heal.go 	modified:   go.mod 	modified:   go.sum 	modified:   internal/agent/agent.go 	modified:   internal/agent/agent_test.go 	modified:   internal/agent/codebase_executor.go 	modified:   internal/agent/cognitive_bus.go 	modified:   internal/agent/cognitive_bus_test.go 	modified:   internal/agent/executors.go 	modified:   internal/agent/llm_mind.go 	modified:   internal/agent/long_term_memory.go 	modified:   internal/agent/models.go 	new file:   internal/agent/models_test.go 	modified:   internal/analysis/active/taint/analyzer.go 	new file:   internal/analysis/active/taint/taint_shim.test.js 	modified:   internal/analysis/active/timeslip/analyzer.go 	modified:   internal/analysis/active/timeslip/analyzer_test.go 	modified:   internal/analysis/active/timeslip/concurrency_test.go 	modified:   internal/analysis/active/timeslip/e2e_test.go 	modified:   internal/analysis/active/timeslip/fingerprint.go 	modified:   internal/analysis/active/timeslip/fingerprint_test.go 	modified:   internal/analysis/active/timeslip/graphql_async.go 	new file:   internal/analysis/active/timeslip/graphql_async_test.go 	modified:   internal/analysis/active/timeslip/h1_concurrent.go 	modified:   internal/analysis/active/timeslip/h1_singlebyte.go 	new file:   internal/analysis/active/timeslip/h1_singlebyte_test.go 	new file:   internal/analysis/active/timeslip/h2_dependency.go 	new file:   internal/analysis/active/timeslip/h2_dependency_test.go 	modified:   internal/analysis/active/timeslip/h2_multiplex.go 	modified:   internal/analysis/active/timeslip/integration_test.go 	modified:   internal/analysis/active/timeslip/oracle.go 	modified:   internal/analysis/active/timeslip/templating.go 	modified:   internal/analysis/active/timeslip/types.go 	new file:   internal/analysis/active/timeslip/types_test.go 	modified:   internal/analysis/auth/ato/analyzer_test.go 	new file:   internal/analysis/static/semanticfp/builder.go 	new file:   internal/analysis/static/semanticfp/canonicalizer.go 	new file:   internal/analysis/static/semanticfp/canonicalizer_test.go 	new file:   internal/analysis/static/semanticfp/coverage_test.go 	new file:   internal/analysis/static/semanticfp/fingerprint_test.go 	new file:   internal/analysis/static/semanticfp/fingerprinter.go 	new file:   internal/analysis/static/semanticfp/helpers_test.go 	new file:   internal/analysis/static/semanticfp/policy.go 	modified:   internal/browser/browser_helper_test.go 	deleted:    internal/browser/dom/config.go 	deleted:    internal/browser/dom/interactor.go 	deleted:    internal/browser/dom/interactor_test.go 	deleted:    internal/browser/dom/xpath.go 	deleted:    internal/browser/dom/xpath_test.go 	modified:   internal/browser/humanoid/behavior.go 	new file:   internal/browser/humanoid/behavior_test.go 	modified:   internal/browser/humanoid/clickmodel.go 	new file:   internal/browser/humanoid/clickmodel_test.go 	new file:   internal/browser/humanoid/drag_test.go 	new file:   internal/browser/humanoid/helpers_test.go 	modified:   internal/browser/humanoid/humanoid.go 	modified:   internal/browser/humanoid/humanoid_integration_test.go 	new file:   internal/browser/humanoid/humanoid_test.go 	modified:   internal/browser/humanoid/interface.go 	modified:   internal/browser/humanoid/keyboard.go 	new file:   internal/browser/humanoid/keyboard_test.go 	new file:   internal/browser/humanoid/mocks_test.go 	new file:   internal/browser/humanoid/movement_test.go 	new file:   internal/browser/humanoid/noise_test.go 	new file:   internal/browser/humanoid/potentialfield_test.go 	modified:   internal/browser/humanoid/scrolling.go 	new file:   internal/browser/humanoid/scrolling.js 	new file:   internal/browser/humanoid/scrolling.test.js 	new file:   internal/browser/humanoid/scrolling_test.go 	modified:   internal/browser/humanoid/trajectory.go 	modified:   internal/browser/humanoid/trajectory_test.go 	new file:   internal/browser/humanoid/vector_test.go 	deleted:    internal/browser/jsbind/dom_bridge.go 	deleted:    internal/browser/jsbind/dom_bridge_test.go 	deleted:    internal/browser/jsbind/errors.go 	deleted:    internal/browser/jsexec/runtime.go 	deleted:    internal/browser/jsexec/runtime_test.go 	deleted:    internal/browser/layout/layout.go 	deleted:    internal/browser/layout/layout_test.go 	modified:   internal/browser/manager.go 	modified:   internal/browser/manager_test.go 	modified:   internal/browser/network/compression.go 	new file:   internal/browser/network/customhttp/client.go 	new file:   internal/browser/network/customhttp/config.go 	new file:   internal/browser/network/customhttp/h1client.go 	new file:   internal/browser/network/customhttp/h2client.go 	modified:   internal/browser/network/dialer.go 	modified:   internal/browser/network/http_parser.go 	modified:   internal/browser/network/http_parser_test.go 	modified:   internal/browser/network/httpclient.go 	modified:   internal/browser/network/httpclient_test.go 	modified:   internal/browser/network/proxy.go 	modified:   internal/browser/network/proxy_test.go 	modified:   internal/browser/network/tls_test.go 	deleted:    internal/browser/parser/css.go 	deleted:    internal/browser/parser/css_test.go 	new file:   internal/browser/session/cdp_executor.go 	new file:   internal/browser/session/cdp_executor_test.go 	new file:   internal/browser/session/context_utils.go 	new file:   internal/browser/session/context_utils_test.go 	modified:   internal/browser/session/harvester.go 	modified:   internal/browser/session/harvester_test.go 	deleted:    internal/browser/session/helpers_test.go 	new file:   internal/browser/session/interaction.go 	new file:   internal/browser/session/interactor.go 	new file:   internal/browser/session/interactor_test.go 	new file:   internal/browser/session/interfaces.go 	deleted:    internal/browser/session/logger_adapter.go 	modified:   internal/browser/session/session.go 	new file:   internal/browser/session/session_helpers_test.go 	modified:   internal/browser/session/session_test.go 	deleted:    internal/browser/session/timing.go 	deleted:    internal/browser/session/timing_test.go 	deleted:    internal/browser/shadowdom/shadow.go 	deleted:    internal/browser/shadowdom/shadow_test.go 	deleted:    internal/browser/shared_types.go 	deleted:    internal/browser/shim/embed.go 	deleted:    internal/browser/shim/shim.go 	deleted:    internal/browser/shim/shim_test.go 	deleted:    internal/browser/shim/taint_shim.js 	modified:   internal/browser/stealth/evasions.js 	modified:   internal/browser/stealth/stealth.go 	modified:   internal/browser/stealth/stealth_test.go 	deleted:    internal/browser/style/engine_test.go 	deleted:    internal/browser/style/selectors_test.go 	deleted:    internal/browser/style/style.go 	deleted:    internal/browser/style/values_test.go 	new file:   internal/mocks/BrowserInteractor.go 	new file:   internal/mocks/BrowserManager.go 	new file:   internal/mocks/KnowledgeGraphClient.go 	new file:   internal/mocks/OASTProvider.go 	new file:   internal/mocks/SessionContext.go 	modified:   internal/mocks/mocks.go 	new file:   internal/service/components.go 	renamed:    cmd/factory.go -> internal/service/factory.go 	new file:   internal/service/initializers.go

+++ internal/browser/humanoid/clickmodel.go
@@ -107,7 +107,8 @@ func (h *Humanoid) calculateTerminalFittsLaw(distance float64) time.Duration {
 		mt = 0
 	}
 
-	return time.Duration(mt) * time.Millisecond
+	// FIX: Convert float64 milliseconds (mt) to time.Duration (int64 nanoseconds) accurately to prevent truncation.
+	return time.Duration(mt * float64(time.Millisecond))
 }
 
 // calculateClickHoldDuration determines how long the mouse button is held down using Ex-Gaussian distribution.
@@ -134,5 +135,6 @@ func (h *Humanoid) calculateClickHoldDuration() time.Duration {
 	// Habituation is already factored into the dynamic Mu/Sigma/Tau.
 	durationMs *= (1.0 + h.fatigueLevel*0.25)
 
-	return time.Duration(durationMs) * time.Millisecond
+	// FIX: Ensure accurate conversion from float64 milliseconds to time.Duration (nanoseconds).
+	return time.Duration(durationMs * float64(time.Millisecond))
 }
